+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 * You can inspect the status of the queue by calling `Ost::Queue#items`.
 
 * If you need access to the underlying Redis key, it's in `Ost::Queue#key`.
+
+* Ost now uses `BRPOPLPUSH` and maintains a backup queue while working.
+
+  You can access this queue using `Ost::Queue#backup`.